+++ app/src/main/java/com/example/movieapp/dagger/component/AppComponent.kt
@@ -2,11 +2,11 @@ package com.example.movieapp.dagger.component
 
 import com.example.movieapp.dagger.module.NetworkModule
 import com.example.movieapp.dagger.module.viewModule.ViewModelModule
-import com.example.movieapp.ui.news.main_news.NewsFragment
-import com.example.movieapp.ui.news.general_news.GeneralNewsFragment
 import com.example.movieapp.ui.home.detail.DetailFragment
 import com.example.movieapp.ui.home.list.ListFragment
 import com.example.movieapp.ui.home.overview.OverviewFragment
+import com.example.movieapp.ui.news.detail_news.DetailNewsFragment
+import com.example.movieapp.ui.news.news.NewsFragment
 import dagger.Component
 import javax.inject.Singleton
 
@@ -16,6 +16,6 @@ interface AppComponent{
     fun inject(overviewFragment: OverviewFragment)
     fun inject(detailFragment: DetailFragment)
     fun inject(listFragment: ListFragment)
-    fun inject(generalNewsFragment: GeneralNewsFragment)
     fun inject(newsFragment: NewsFragment)
+    fun inject(detailNewsFragment: DetailNewsFragment)
 }
