+++ app/src/main/java/com/example/movieapp/dagger/module/viewModule/ViewModelModule.kt
@@ -3,9 +3,7 @@ package com.example.movieapp.dagger.module.viewModule
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
 import com.example.movieapp.ui.home.detail.DetailViewModel
-import com.example.movieapp.ui.news.main_news.NewsViewModel
 import com.example.movieapp.ui.home.list.ListViewModel
-import com.example.movieapp.ui.news.general_news.GeneralNewsViewModel
 import com.example.movieapp.ui.home.overview.OverviewViewModel
 import com.example.movieapp.utils.ViewModelKey
 import dagger.Binds
@@ -37,8 +35,8 @@ abstract class ViewModelModule {
 
     @Binds
     @IntoMap
-    @ViewModelKey(GeneralNewsViewModel::class)
-    internal abstract fun bindGeneralNewsViewModel(viewModel: GeneralNewsViewModel): ViewModel
+    @ViewModelKey(NewsViewModel::class)
+    internal abstract fun bindGeneralNewsViewModel(viewModel: com.example.movieapp.ui.news.news.NewsViewModel): ViewModel
 
     @Binds
     internal abstract fun bindViewModelFactory(factory: ViewModelFactory): ViewModelProvider.Factory
