+++ app/src/main/java/com/example/movieapp/ui/search/search/SearchFragment.kt
@@ -14,7 +14,7 @@ class SearchFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.main_search_fragment, container, false)
+        return inflater.inflate(R.layout.search_fragment, container, false)
     }
