+++ app/src/main/res/layout/detail_news_fragment.xml
@@ -1,0 +1,10 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools">
+    <androidx.constraintlayout.widget.ConstraintLayout
+       android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        tools:context="ui.news.detail_news.DetailNewsFragment">
+
+    </androidx.constraintlayout.widget.ConstraintLayout>
+</layout>
