+++ app/src/main/res/layout/main_home_fragment.xml
@@ -19,7 +19,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:navGraph="@navigation/main_graph" />
+            app:navGraph="@navigation/home" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
 </layout>
