+++ app/src/main/res/navigation/search.xml
@@ -4,9 +4,9 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:id="@+id/search"
     app:startDestination="@id/searchFragment">
-    <fragment
+    <keep_state_fragment
         android:id="@+id/searchFragment"
         android:name="com.example.movieapp.ui.search.search.SearchFragment"
-        android:label="SearchFragment"
-        tools:layout="@layout/main_search_fragment"/>
+        android:label="search_fragment"
+        tools:layout="@layout/search_fragment" />
 </navigation>
